+++ ChangeLog.md
@@ -1,7 +1,18 @@
+## v1.3.9
+
+- discord.js, @discordjs/voice version bump
+- /nowPlaying now includes thumbnails
+- Added command for toggling PlayerDebugState
+- Moved GuildPlay to PlayerHandler
+- Adding (missing) NOTICE handling to EventHandler
+- Covering event GuildAuditEntryCreate from AuditHandler
+
+#
+
 ## v1.3.8
 
 - discord-player, equalizer & extractor version bump
-- Switched to discord-player usePlayer() 
+- Switched to discord-player usePlayer()
 - Added ffmpeg-static to fix early track end bug
 
 #